requirements.txt
feedparser
requests
python-dotenv
openai
beautifulsoup4
lxml
schedule
pytz
